go_learn/imooc-go-learn/spider/single-task/zhenai/parser/citylist/citylist.go
package citylist

import (
	"imooc-go-learn/spider/single-task/engine"
	"regexp"
)

const cityListRe = `<a href="(http://www.zhenai.com/zhenghun/[0-9a-z]+)" [^>]*>([^<]+)</a>`

func ParserCityList(contents []byte) engine.ParseResult {
	re := regexp.MustCompile(cityListRe)
	matches := re.FindAllSubmatch(contents, -1)

	result := engine.ParseResult{}
	for _, m := range matches {
		result.Items = append(result.Items, string(m[2]))
		result.Requests = append(result.Requests, engine.Request{
			Url: string(m[1]),
			ParserFun: engine.NilParser,
		})
		//fmt.Printf("city: %s, url: %s\n", m[2], m[1])
		//fmt.Println()
	}
	return result
}
